--- data/check.md5
@@ -1,5 +1,0 @@
-eced898f45e840f61b8d973893ba491c  enwik3.raw
-0f40d55c4ef1958bb4ab24f5c32231f4  enwik4.raw
-1a80469f183a11d050e37e908bc65f75  enwik5.raw
-83a2cefd8ca00049e0445542dabd25a0  enwik6.raw
-2d1946513a5e933b897180084549dfc0  enwik7.raw
