--- data/check.sha1
@@ -1,5 +1,0 @@
-f90b7108ee1396f0bd6f1d20c17f802f1a2d0a9c  enwik3.raw
-52b1798288dcb6a4c2398a969d3b55b31dc5991c  enwik4.raw
-b8a7421a36d13710e379f05badbf0e00c48918af  enwik5.raw
-d89a92c68c1346a6dffbd62be1b14c411dabcfd7  enwik6.raw
-704554c77aa78e36e10cdf82092b84e3e0cbea85  enwik7.raw
