--- data/textual/enwik3.raw
@@ -1,20 +1,0 @@
-<mediawiki xmlns="http://www.mediawiki.org/xml/export-0.3/" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.mediawiki.org/xml/export-0.3/ http://www.mediawiki.org/xml/export-0.3.xsd" version="0.3" xml:lang="en">
-  <siteinfo>
-    <sitename>Wikipedia</sitename>
-    <base>http://en.wikipedia.org/wiki/Main_Page</base>
-    <generator>MediaWiki 1.6alpha</generator>
-    <case>first-letter</case>
-      <namespaces>
-      <namespace key="-2">Media</namespace>
-      <namespace key="-1">Special</namespace>
-      <namespace key="0" />
-      <namespace key="1">Talk</namespace>
-      <namespace key="2">User</namespace>
-      <namespace key="3">User talk</namespace>
-      <namespace key="4">Wikipedia</namespace>
-      <namespace key="5">Wikipedia talk</namespace>
-      <namespace key="6">Image</namespace>
-      <namespace key="7">Image talk</namespace>
-      <namespace key="8">MediaWiki</namespace>
-      <namespace key="9">MediaWiki talk</namespace>
-      <namespa
