+++ testdata/check.md5
@@ -1,0 +1,3 @@
+2d1946513a5e933b897180084549dfc0  enwik7.raw
+10a9447b8d1849480bdcbab20023f8eb  f32_256normal7.raw
+b14f131dda5bae6d928fa249525ed070  u32_uniform7.raw
