+++ testdata/check.sha1
@@ -1,0 +1,3 @@
+704554c77aa78e36e10cdf82092b84e3e0cbea85  enwik7.raw
+f7fc2c6b4c6305d09d82a1eb07a962603bdcae1b  f32_256normal7.raw
+e9520ea5d18ebf58bc0c87f0e2e55465bf77fc22  u32_uniform7.raw
